Lab-01/requirement.txt
dicttoxml==1.7.16
mysql-connector-python==9.0.0
PyMySQL==1.1.1
